requirements.txt
apache-airflow==2.10.4
pandas==1.5.3
requests==2.31.0
boto3==1.28.30